end work with people

+++ PSS_HomwWork2/Additional_classes/date.cpp
@@ -4,9 +4,9 @@ Date::Date(unsigned short d, unsigned short m, unsigned short y)
     :year(y), month(m), day(d)
 {}
 
-int Date::getAgeDiff(const Date &r) const
+unsigned short Date::getAgeDiff(const Date &r) const
 {
-    int diff = year - r.year - ((r.month > month)?1:((r.month == month)&&(r.day > day)));
+    short diff = year - r.year - ((r.month > month)?1:((r.month == month)&&(r.day > day)));
     if(diff<0)
         diff = r.year - year - ((month > r.month)?1:((month == r.month)&&(day > r.day)));
     return diff;
